+++ src/main/java/dev/masa/masuitecore/bungee/listeners/LoginEvent.java
@@ -1,4 +1,4 @@
-package dev.masa.masuitecore.bungee.events;
+package dev.masa.masuitecore.bungee.listeners;
 
 import dev.masa.masuitecore.bungee.MaSuiteCore;
 import dev.masa.masuitecore.core.channels.BungeePluginChannel;
@@ -44,11 +44,11 @@ public class LoginEvent implements Listener {
                     ServerInfo serverInfo = entry.getValue();
                     serverInfo.ping((result, error) -> {
                         if (error == null) {
-                            new BungeePluginChannel(plugin, serverInfo, new Object[]{
+                            new BungeePluginChannel(plugin, serverInfo,
                                     "MaSuiteCore",
                                     "AddPlayer",
                                     e.getPlayer().getName()
-                            }).send();
+                            ).send();
                         }
                     });
                 }
